virt: Wait for event rather than checking domain XML on hot unplug
When a device is hot unplugged synchronously, _waitForDeviceRemoval asks
libvirt for domain XML periodically and checks the device presence there.  Now,
when we have device removal libvirt event, we can replace that with waiting for
the event.

An exception is lease device, which doesn't generate a device removal event on
hot unplug in libvirt.  We have to keep examining lease device removal by
checking domain XML changes.

Change-Id: I258b45bfb19606eb68effce5f9890bb9d37be621
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>