virt: Unify synchronous and asynchronous device removal
There are common device removal actions that are duplicated for synchronous and
asynchronous device hot unplugs.  The other problem is that if a "synchronous"
device hot unplug finishes only after its timeout, device removal and cleanup
are not performed.  So let's handle all device removals in the device removal
callback.

We no longer attempt to prevent parallel hot unplugs of the same device by
"premature" device removal as hot unplugs of *some* devices did.  The reason is
that a misleading message about device not found is not the best solution, it's
better to let the hot unplug simply pass through the libvirt call.  Engine
shouldn't permit duplicate hot unplugs with the exception of timed out
synchronous hot unplugs.  Then the libvirt call can either fail or succeed on
the duplicate attempt, in the latter case the device cleanup will be performed
only once, in onDeviceRemoved callback.  We are safe in any case, without
adding code complexity.

One significant drawback of this patch is that we call device teardown and
domain descriptor update from libvirt in a libvirt callback now.

Change-Id: Ie63db0439ae4f410ce11ac194d7bdd055650b4c3
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>