tests: Reproducer for bug 1553133
In lvm2 < 2.03, lvm commands such as lvs, vgs, pvs, and "lvchange
--refresh" running on non-spm host may corrupt vg metadata.

This is a stress test reproducing the issue on lvm2 < 2.03 (RHEL 7.7),
and verifying that the issue was fixed on lvm2 >= 2.03 (RHEL 8.1).

To reproduce the issue on lvm2 < 2.03, or verify that the issue was
fixed with lvm2 >= 2.03, run the regular node without the --read-only
option.

To test the fix on lvm2 < 2.03 using locking_type=4, run the regular
node with the --read-only option.

The script provides 4 sub commands:
- run-manager
- run-regular
- create-vg
- remove-vg
- log-stats

The run-manager command starts now a simple socket server using JSON
line based protocol.

The run-regular command starts multiple workers based on --concurrency
option. Every worker connect to the server and perform this flow:

    for every iteration:
        create LV
        activate LV
        repeat 20 times:
            extend LV
            refresh LV
            sleep 0-2 seconds
        deactivate LV
        remove LV

The read-write operations (create LV, extend LV, remove LV) are
performed on the manager node. The read-only operations are performed
locally.

Read-only commands are performed using locking_type=1 with --read-only
is not set, and locking_type=4 when --read-only is set. locking_type=4
ensures that lvm does not attempt to recover the metadata when the
metadata header was modified during the read. If a command find the
storage in inconsistent state, it will fail instead of attempting to
recover the metadata, which is likely to corrupt the VG metadata.

To handle the expected failures of read-only commands, we use a retry
mechanism with exponential back-off. When a command fails, we wait and
retry the command again; each time the command fail we double the wait
time to let the manager finish pending metadata updates.

I tested creating 50,000 LVs, and then extending and refreshing them 20
time, total of 1,000,000 successful operations at rate of 29.7
operations per second.

I also did some short runs creating 500 LVs, to check retry failure rate
with lower operation rates.

Here is a table with the results from all runs:

extend-delay  extends  retries  max-retry  total-time  extend-rate  retry-rate
------------------------------------------------------------------------------
           1  1000000   237250          8       33677        29.69      21.56%
           2    10000     1027          5         491        20.36       9.33%
           5    10000      221          2        1138         8.78       2.00%
          10    10000      100          2        2220         4.50       0.90%
          20    10000       35          1        4382         2.28       0.31%
          40    10000        8          1        8585         1.16       0.09%

Change-Id: I192ebdf75b4fb53298188c9913712044277f8525
Related-To: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>