virt: Set correct owner of oVirt guest agent socket
Current libvirt sets proper owner and group of guest agent sockets only if
dynamic ownership is enabled.  Without dynamic ownership the owner and group
are set to root.  It's not a problem for QEMU guest agent because it is handled
by libvirtd running with root permissions.  But oVirt guest agent is
inaccessible to Vdsm unless dynamic ownership is enabled.

oVirt 4.2 doesn't enable dynamic ownership so we must set the proper socket
owner and group before accessing the socket.

Change-Id: I787ad0acbad17a75916daba262f5032405be4088
Bug-Url: https://bugzilla.redhat.com/1635687
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>