deprecated: Mark deprecated infrastructure
Both AsyncProc and AsyncProcessOperation should not be used in new code.

AsyncProc is not compatible with python 3 and highly inefficient.
Callers that need a timeout can use
vdsm.compat.subprocess.Popen.communitate(timeout=timeout).

Callers that need to stream process stdout should use
cmdutils.receive().

AsyncProcessOperation is using AsyncProc to to make it easy to run a
process with a timeout. This is usually the wrong thing and we should
not make it easy.

Change-Id: I043bf4d89af5ba31b0457941c8c2718f6b4071bc
Signed-off-by: Nir Soffer <nsoffer@redhat.com>