gluster: Fix errors related to heal info command
vdsm-gluster tries to run heal operations on all
volumes.Gluster commands may get stuck and then
vdsm starts waiting for them indefinetly, thus exausting
threads.This causes vdsm to timeout and stop
communication with the engine and will go offline.

Adding a timeout to the gluster heal info command
so that threads are not exhausted and the command
terminates within set timeout.On encountering a
timeout the system will issue GlusterCommandTimeoutException
which will cause the command to exit and notify engine
of the state.

Change-Id: I565c7154f51dbf88c3f3d8b64f1997a139e2dcab
Bug-Url: https://bugzilla.redhat.com/1614430
Signed-off-by: Kaustav Majumder <kmajumde@redhat.com>