vm: livemerge: sync self.conf devices disk data
Make sure the data we store in self.conf['devices'] for disks
is up to date when live merge completes.

This is needed when Vdsm >= 4.2 is working in backword compatibility
mode. We still store a copy of the disk data in self.conf['devices']
to support the livemerge flows and the getVMFullList API.

Change-Id: Ia693cf558723338bb3cf557275fcc8532ef6dd2c
Bug-Url: https://bugzilla.redhat.com/1623467
Topic: livemerge-legacy-conf-update
Signed-off-by: Francesco Romani <fromani@redhat.com>