commands: Introduce 'run' and 'start' functions
'run' - replaces 'execCmd()'. Starts a child process, communicates
with it, and waits until the child process terminates. Returns
the command output or raises if the child process terminated
with non-zero exit code.

'start' - replaces 'execCmd(sync=False)'. Starts a child process -
the caller is responsible for communicating with the subprocess
and waiting for it. This is needed by code that want to write
lot of data to the subprocess stdin, or read data from the child
process stdout and stderr, without buffering the data.

Changes compared with 'execCmd':
- Raises 'cmdutils.Error' including all the information about
  the failing command.
- Prevents logging arguments wrapped in 'ProtectedPassword'.
- Uses 'sudo' to kill commands only when they were run with 'sudo'.
- Possibility to run processes asynchronously in Python 3
  ('execCmd(sync=False)' was Python 3-incompatible)

Based on: https://gerrit.ovirt.org/83923/2

Change-Id: I0c81256dcdbd5fbd685028f0cd8dd432d6fd21a5
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>