net: py3 compatibility - Do not use None for comparison
On py3,'None' cannot be compared (except equality) to
other objects, and therefore cannot be passed as an
argument to the max func.

Normalize possible 'None' argument to 0 when passing to
the max func.

Verified by running the following on fedora with py3:
pytest --target-lib network/functional/bond_basic_test.py

Change-Id: I279a17b364881bfb0381f86943dd11a1e09a9d34
Signed-off-by: Bell Levin <blevin@redhat.com>