virt: Support for vGPU placement
vGPU devices are currently placed at the first available card.  The user may
want to place vGPU devices on separate cards, e.g. for performance reasons.

This patch introduces two vGPU placement policies, `compact' to use the first
available card as till now, and `separate' to prefer using separate physical
devices if possible.  The current policy remains default for compatibility.

vGPU placement is specified by attaching the placement policy name to mdev_type
metadata using `|' character.  Selection of the given policy for the given vGPU
is made by Engine.

It is assumed that the placement policy is and remains the same for all vGPUs on the
given host.  This patch doesn't handle policy changes.  For instance, if a vGPU
is assigned with separate placement and another vGPU is assigned with compact
placement, nothing prevents assigning the second vGPU on the same card as the
first vGPU.  If we need to prevent that, it'll be handled in a separate patch.

Change-Id: I9d4ad5baa987f5758b94e69751ae7cc3d70e4330
Bug-Url: https://bugzilla.redhat.com/1641125
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>