storage: add os-brick helper for caps
Adding OS brick connector information to getCapabilities.

A new entry named 'connector_info' is added.

This information will be used by the Engine in order to
expose a disk to this host using CinderLib.

Note that currently os-brick package it not required in the
Vdsm spec as it is not available in oVirt channels.
Therefore, the presence of the package is tested before trying
to invoke it.
If os-brick is not available, the new entry will not be available.

Here an example of the connector information:

"connector_info": {
        "initiator": "iqn.1994-05.com.redhat:3f2b2be7ebc",
        "ip": null,
        "platform": "x86_64",
        "host": "vdsm-dev",
        "do_local_attach": false,
        "os_type": "linux2",
        "multipath": true
    }

Change-Id: Id9efa157e5f67d41839b4e05a7849efc33c8f6f3
Signed-off-by: Fred Rolland <frolland@redhat.com>