storage: add attach_volume verb
Attach a device using os-brick according to the connection information
received.

The verb will return the device information after it was attached.
Using this information, the Engine will update the VM XML before
running it.

Note that if os-brick is not installed, an ManagedVolumeNotSupported
error will be returned.

In order to be able to detach the volume, the volume information
is saved in a DB.
This DB can be used to filter out the attached multipath devices
from the other Vdsm flows.

Change-Id: I07abfcb00e39ed951b7b91ff3e0d72feabd21f9a
Signed-off-by: Fred Rolland <frolland@redhat.com>