build: Cleanup 'PyYAML' dependency declarations
'PyYAML' is a build dependency for 'vdsm-api' and a runtime dependency
for 'vdsm-gluster'. Please note, that nothing new is added - this patch
only moves the dependency declaration from 'vdsm' package level
to proper packages that really use 'PyYAML'.

An important result of this movement is that 'vdsm' installation
without 'gluster' packages doesn't require 'PyYAML' as a runtime
dependency at all.

Change-Id: I9c7572e413ccde0c6b67a85500282552c3cd00df
Bug-Url: https://bugzilla.redhat.com/1651552
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>