qemuio: Support qemu 2.12 error handling
qemu-io 2.10 was exiting with zero exit code and "Pattern verification"
error in stdout. In 2.12, non-zero code is returned when pattern
verification fails. This sounds better, but the exit code is not
specific enough to detect the verification error.

Check for pattern verification error before testing exit code, so we
raise (expected) verification error in this case, and raise a generic
error in verification error was not detected.

Change-Id: Id108d2b780b6862bed3bb7088908d911788b544d
Signed-off-by: Nir Soffer <nsoffer@redhat.com>