mount: Workaround "target is busy" in mount tests
Fedora 28 mount tests fail randomly when unmounting with this error:

    vdsm.storage.mount.MountError: (32, b';umount:
    /tmp/vdsm-mkimage-teststvi1s2gj/work: target is busy.\n')

This looks like the same issue we had 2 years ago with older Fedora.
We fixed this in the past in commit d158b1a6f54b (mount: Wait for
systemd temporary mount unit) by waiting for udev events after running
mount and umount.

With real storage we never umount immediately after mount, so I think
this is a test issue, and we can work around it using a small delay
before umounting.

Change-Id: I06ac979b6400eab3b4f953feafccd810fd12c50f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>