qga: report file system stats and disk mappings
Disk mapping is not tracked as part of guest-info but is kept separate.
To avoid pulling yet another variable to/from the poller we store the
mapping in guest-info in poller and move it into the proper place in
getGuestInfo(). As usual the content from oVirt guest agent takes
precedence.

Change-Id: I263110227a5d3303b1736ebed41bd93345ebc5a3
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>