ci: Remove 'index.html'
I find 'index.html' useless. On 'build-artifacts' the only thing that
works is 'vdsm-api.html' link. On 'check-patch' it is included
conditionally (only if we build packages) and our 'check-merged'
is a subject for refactorization. I vote for removing it. If we need
something similar we should work on a better option after splitting
the CI to substages.

Change-Id: If3dd0b32026ff875e58e161256ec2da936a48a2b
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>