storage: support attachment of rbd devices to VMs
Since rbd paths are not handled by multipath we need
to handle their preparation separately, hence changing the
ownership of each device so it can be accessed by qemu.

For this purpose the appropriateMultipathDevice was changed
to accomodate RBD devices as well. However, since we do not
have a WWN like identifier for RBD devices we send the entire
device.

Change-Id: I177233fded9f68c61bbab4fbc62ff7001335026b
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>