supervdsm: Add command line options for testing
When testing code depending on having running supervdsm, we want to run
supervdsm as root on a developer machine, without /etc/vdsm and
/var/log/vdsm.

Add new command line options allowing running supervdsm:
- serving $USER:$USER instead of vdsm:kvm
- with custom logger conf, logging to stderr.
- disabling gluster
- disabling network initialization

A sample svdsm.logger.conf was added to tests directory to use with
--logger-conf.

Here is an example session showing how these options can be used:

$ sudo mkdir /run/vdsm

$ sudo chown $USER:$USER /run/vdsm

$ sudo env PYTHONPATH=lib static/usr/sbin/supervdsmd \
      --sockfile /run/vdsm/svdsm.sock \
      --sock-user=$USER \
      --sock-group=$USER \
      --logger-conf tests/conf/svdsm.logger.conf \
      --disable-gluster \
      --disable-network

At this point you can run tests that depend on a running supervdsm as a
normal user, on your laptop. We will use this for testing the new NBD
APIs.

Change-Id: Ia88e331af43e5ad21a1ab3d8050de89a65e6f190
Signed-off-by: Nir Soffer <nsoffer@redhat.com>