clientIF: Fix reactor life cycle
Previously we created the reactor when initializing clientIF, and
started the reactor thread in clientIF.start(). However, the reactor was
closed from the acceptor. This works since we stop both the acceptor and
reactor during shutdown, but it is wrong design, leftover from the time
the reactor was created by the acceptor.

Things should be started and stopped in the same place. If clientIF is
starting the reactor, it should also stop it. The reactor cannot depend
on the acceptor using it.

This patch fixes the wrong dependency by stopping the reactor in
clientIF.prepareForShutdown.

Change-Id: I319d06e8cffd86f87417053137f94eab90a4998d
Signed-off-by: Nir Soffer <nsoffer@redhat.com>