net: Control how the sourceroute api is called
During vdsmd initialization, vdsmd calls the sourceroute
initialization through supervdsmd.

The initialization step should not be dependent on a running
supervdsm service, allowing it to be decoupled from the way VDSM
is deployed and specifically, allowing it to be covered by tests.

Following the patch, the initialization now depends on the network
API interface, and can use either a given supervdsm proxy or the
direct network api module.

Change-Id: Ib92db9c28d10f9ddb5cfe771b73869b5b1e70a76
Signed-off-by: Bell Levin <blevin@redhat.com>