spec, configure.ac, build: refactor build stage determine configure flags in configure.ac
In this patch, we determine the flags to configure inside configure.ac and insert them into the spec file
when it is being generated, this makes the spec file easier to understand, and cleaner.

When we will build for python2/3 we need to run configure and make in separate folders
(this is done in multiple projects and is the recommended way to run configure with different
python interpreters) This patch will save us code duplications in the upcoming patches.

Change-Id: Ifaaed8c06c531a3d71dde7a2c7f683f3f56b7ab0
Signed-off-by: gzaidman <gzaidman@redhat.com>