systemd: Allow systemctl.stop() without supervdsm
Change supervdsm systemctl_stop so it can be run directly without
running supervdsm if the process is running as root. This makes it
simpler to test code depending on supervdsm.

This hides the supervdsm machinery from callers. For example, we can do
now:

    from vdsm.common import systemctl

    systemd.stop("vdsm-nbd-985fb85c-3b86-4635-b12a-8eba70ef504f")

If we run as root, the command is invoked directly. Otherwise this calls
supervdsm systectcl_stop(), which calls systemctl.stop() with the given
arguments.

Finally, this simplifies the supervdsm_api code, since the actual
implementation is inside the systemd module.

Change-Id: Ib731cabdc8e057698b407e69b0609db8719db522
Signed-off-by: Nir Soffer <nsoffer@redhat.com>