systemd: Allow systemd_run() without supervdsm
Like systemctl_stop(), if we run as root, we can invoke now
systemd_run() without running supervdsm. This can be useful for testing
code that depends on supervdsm without actually running supervdsm.

This hides the supervdsm machinery from callers:

    from vdsm.common import systemd

    systemd.run(["true"])

The module provides also wrap() function, for code that already running
as root, and need to invoke a command with systemd-run only in some
cases:

    if foobar:
        cmd = systemd.wrap(cmd)

    return commands.run(cmd)

Change-Id: I691599e11301d194403fb6e0d48053e5dc157023
Signed-off-by: Nir Soffer <nsoffer@redhat.com>