travis: Use shallower clone
We used depth=999999 to ensure that we have a tagged commit. Without a
tagged commit, autogen.sh fail to generate a package version.

Change depth to 1 for quicker clones, and add a fake v0 tag to keep
autogen.sh happy.

This save about 7 seconds.

Here is git command before this change:

$ git clone --depth=9999999 --branch=master https://github.com/nirs/vdsm.git nirs/vdsm
8.18s

And after this change:

$ git clone --depth=1 --branch=travis https://github.com/nirs/vdsm.git nirs/vdsm
1.17s

Change-Id: I24963ef97e492736ea8486085c4c22dc1006af92
Signed-off-by: Nir Soffer <nsoffer@redhat.com>