build: Ignore tar errors when collecting logs
Tar fail randomly when some log file is modified during collection, and
there is no way to disable this behavior or detect this error. We ignore
now any error when collecting logs to avoid this issue.

When tar fail, we sill see the error from tar, and a warning like:

    WARNING: Ignoring error collecting logs in /var/log

Change-Id: Ia5f471e8588ae26007ed4a1cef1365620ec49d56
Signed-off-by: Nir Soffer <nsoffer@redhat.com>