net tests: remove leftover bond for tests to pass
following a new test that was added [1], the bond's failure
to be erased was not taken into account. The erasing occurs as
part of the setupNetwork cleanup, but since the setup fails
(as expected), the cleanup never occurs.

Fixed by explicitly deleting the bond, to ensure the bond does
not linger and interfere when preforming a series of
sequencial tests.

[1] https://gerrit.ovirt.org/#/c/95784/

Change-Id: I06825e4514ad7255db19786a49a142a13e950f82
Signed-off-by: Bell Levin <blevin@redhat.com>