net: Introduce the nmstate network provider
Enable the creation of networks and bonds using nmstate.

The current implementation is limited in scope, covering:
- Linux bonds
- Bridgeless networks
- Static and dynamic IPv4, IPv6 (missing static routes)
- VLANs

The reporting is still performed through iproute2, sysfs and
netlink.
The original rollback support is used for the check connectivity stage,
nmstate internal rollback is used for all the rest (errors in the
transaction or an incorrect state outcome).
VDSM persistency (running and persisted) is still used to support the
detection of the logical network entities and allow migrating back from
nmstate.

The option to use nmstate as a provider is controlled from vdsm.config
using the `net_nmstate_enabled` parameter.

Change-Id: If7dd921965343ec99a8eb24a437123e8f9f93e08
Signed-off-by: Edward Haas <edwardh@redhat.com>