virt: Set device hotunplug_event only after device teardown
Currently, device hot unplug event is set and thus device hot unplug
call finishes as soon as successful hot unplug is reported from
libvirt.  But the unplugged device may not yet be available at the
moment for further use, this is guaranteed only after device
teardown.  For instance, NIC SR-IOV instances must be returned to the
system after hot unplug.

Once a device hot unplug finishes, Engine refreshes host capabilities.
If the device teardown is not yet finished, the device is reported to
Engine as unavailable and Engine handles it as such, until the next
automatic or manual host capabilities refresh.

This patch fixes the problem by setting device hotunplug event only
after device teardown.

It's another fix, in addition and unrelated to
https://gerrit.ovirt.org/96092, needed to fix NIC hot unplug
regressions reported in the referred bug.

Change-Id: Ifcc74d2eef62a40ce6f6c924a0eb2003e4b6b5df
Bug-Url: https://bugzilla.redhat.com/1655276
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>