tests: Test building and updating command config
Add test for LVMCache._addExtraCfg(), adding the --config argument to
lvm commands. The current code lookup all devices using multipath,
generate a filter, and cache the config. The same config is used until
the filter is invalidated.

To control locking_type in the config, we need to cache the filter or
the devices instead of the config string, so when we change read only
mode, we don't have to lookup the multipath devices again.

Add tests verifying the current code before we change caching mechanism.

Change-Id: Ie04fd376e60557ee82108bf7f1f77db5716e1369
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>