schema: Fix blockSize default value
Since commit 1d8f35cbdad6 (storage: Added 4k block support API flags),
creating a storage domain fails with:

    InvalidParameterException: Invalid parameter: 'block_size=None'

The default was set to null, meaning vdsm will auto-detect the block
size. However auto-detecting block size is not implemented yet, and even
when it will be, it will work only for GlusterFS, so this is not a good
default.

Change the default to 512 so if engine does not specify a value we use
the V4 behavior.

Change-Id: I5f9744b23302adfadf16ecf40926ab0c3fd34cc1
Signed-off-by: Nir Soffer <nsoffer@redhat.com>