automation: Create more loop devices
When we start the build in Jenkins, we typically find 5 loop devices,
and add 3 devices. Even if the 5 devices were busy, the 3 new devices
should be enough since we use only one device at the same time. However
the tests fail randomly (once per few weeks) with this error:

    MountError: (32, ';mount:
    /tmp/vdsm-mkimage-testsxtGqen/images/vmId_floppy.e83b907eacb704d3942b8cab79561d2a.img:
    failed to setup loop device: No such file or directory\n')

Increase the number of loop devices for testing if this hides this
issue. When we have more time we should actually understand this failure
and apply a real fix.

Change-Id: I6b77bbbfca31c0a9d200ac9ce06b351e2053928b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>