lvm: Retry read-only commands
When running read-only commands, we expect the command to fail if the
SPM modified VG metadata while the command was reading it.

Unfortunately we don't have a way to detect this specific error, so we
must retry on any failure. Maybe we can optimize this later by aborting
the retry if we detect unexpected error in the command stderr.

Add a third step to existing retry mechanism, retrying read only
commands with exponential back-off. When a read-only command fail, we
wait 10 milliseconds before retrying the command again, and double the
time on every failure, up to 10 seconds delay for last retry.

Testing with extreme load show that the first retry is expected to
succeed so we log failed retries as warnings.

Change-Id: I1dfd42240010198456239693d925bf235795529f
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>