lvm: Limit number of concurrent commands
Since 3.6 we fixed locking issue that most of the time serialized lvm
cache refreshes. This fix creates unwanted I/O when when using lot of
storage domains.

This limit is also wanted for the read-only retry mechanism. This may
help to decrease the load on the SPM when read-only commands are
retrying.

Add un-configurable 10 commands limit, so we don't have to test multiple
settings.

Change-Id: I04ac66402120ff22548ffaf57c54029f40911888
Bug-Url: https://bugzilla.redhat.com/1553133
Bug-Url: https://bugzilla.redhat.com/1658866
Signed-off-by: Nir Soffer <nsoffer@redhat.com>