lvm: Support checking block size of loop device
When checking block size of a device, we assume that the device is a
multipath device, and use multipath.getDmId() to find the device name.
Change the code to use os.path.realpath(), so it can work both with
multipath devices (e.g. /dev/mapper/xxxyyyy -> /dev/dm-12) and loop
devices (e.g. /dev/loop3).

Change-Id: I91bad606e49a30e623feb9259271ba04d08e52d4
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>