stdci: Pin slaves to the same mock distribution for tests
The CI jobs run VDSM tests with specified distributions on slaves which
may operate with a totaly different distribution.
Such scenarios may cause userspace programs to run with a non supported
kernel (e.g. fc28 mock running on a el7 slave) and fail randomly.

With this change, the mock and slave host are sync to use the same
distribution.

Change-Id: I958c8b3e3329d20618411444a71ecbc3b08a4d82
Signed-off-by: Edward Haas <edwardh@redhat.com>