tests: Test retrying commands on errors
Test the retry mechanism for both read-write and read-only commands.

To speed up testing, we mock the retry delay and misc.execCmd so we
don't run any lvm command. Running lvm commands is covered by the
functional tests.

The new tests duplicate some of the build commands and build filter
tests. We will replace these tests later with some more command test, so
we don't test private methods.

This change increases coverage to 72% with python 2.7 and 49% with
python 3.6.

Change-Id: I75735e3066747fbc01e8e5d08467cc29b74c8fbd
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>