lvm: Add lvm.set_read_only()
The new function is used in the tests to switch lvm module to read-only
mode. It will be used soon in StoragePool to switch to read-write mode
only when the host is the SPM.

The tests revealed an issue with retrying read-only commands. Commands
like:

    vgs vg-name --config 'global { locking_type=4 }'

Fail when vg-name does not exists, and our retry mechanism try the
failing commands 10 times which takes about 15 seconds. This is not a
new issue, but before we had only one retry, and now we have 11.

Until this issue is solved, disable read-only mode when checking if a VG
was removed.

Change-Id: I7b50d2522c4534966bdf7e17644e158bbe6c0c45
Bug-Url: https://bugzilla.redhat.com/1553133
Signed-off-by: Nir Soffer <nsoffer@redhat.com>