automation: Require same host and distro for tests
Since we started to require same host and distro for any build, Jenkins
queue is overloaded and build take hours instead of minutes, because we
use Fedora 28 for most builds, and it seems that the CI does not have
enough Fedora 28 slaves.

We need same host an distro only for the tests, in particular the
network tests. We don't need it for the linters run, or for the install
check. Move the requirement to the tests sub stages.

Change-Id: I4587c5e8632dda7404aab8d5d1e5717dc593db6d
Signed-off-by: Nir Soffer <nsoffer@redhat.com>