managedvolume: Detach if _resolve_path() failed
Move _resolve_path() call into the try block, so on failure we detach
the volume. Fix the code and the test verifying the wrong behaviour.

Also add missing check that we don't leave the volume in the db in this
case, and add comments to make the test more clear.

Change-Id: I841425ee04fce95948a4b7e5b2a801aab2597d7c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>