spec: Use wildcards for virt files
There is no need to list each virt source file individually in
vdsm.spec, let's use wildcards instead.

The change requires to move the only exception, vm_migrate_hook.py, to
a separate directory.  Which is good anyway, because that file is
actually not a part of vdsm.virt package.

Change-Id: I6d303efa5e1f86d5d028720c68236ab1172a0341
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>