managedvolume: Invalidate lvm filter after attach
After successful attach, we store the multipath id in database, so
the next time lvm filter will be created will not include the managed
volume. However if vdsm already discovered the new device after the
device was connected to the host on the storage server (FC), or after
the device was attached but before the multipath id was stored, it may
be included in the lvm filter.

Invalidated lvm filter after attach to make sure the device is not
included in the lvm filter.

Change-Id: I2cd3bf58d4d57c11f99ce64afeab1797d375ec8a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>