virt: In FIPS mode, use SASL auth instead of qemu passwords
When a host is running in FIPS mode, SASL authorization must be used
instead of qemu password. This patch modifies
Vm._setTicketForGraphicDev accordingly, as well as adds support
functions in supervdsm.

Also, when starting a VM, configuration must be checked to disallow
insecure condition: the engine in FIPS mode won't send a password but if
SASL is not set up, qemu would allow to connect without any
credentitals. graphics.isVncSecure function checks for this condition.

The ticket is created exactly as for regular VNC authentication. The
engine only marks it with FIPS-mode when the host has FIPS line in
kernel parameters.

Change-Id: I0bdbb718da567c9dd9616b6f2bb4b1bbc7fe9ba2
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1695567
Signed-off-by: Tomasz Baranski <tbaransk@redhat.com>