blockSD: Simplify getting occupied metadata slots
The current code supports variable number of blocks for volume metadata.
This is a horrible idea that makes it harder to allocate, read, and
write metadata, and was never used.

This code was added in commit 496c0c37df38 (BZ#732980 MD tag must be in
blocks unit). The code tried to support 4K drives by allocating 1 block
(512 or 4096) for volume metadata, based on the storage block size.

We want to have a unified format for both 512 and 4k drives, with
constant size (8K) per volume metadata, so the code for allocating,
reading, and writing volume metadata can be the same regardless of the
block size.

Change _getOccupiedMetadataSlots() to stop reading the
TAG_PREFIX_MDNUMBLKS tag (e.g. "MS_1") which was never written to
storage.

This does not change the behavior of the system, since we always used 1
block.

To minimize this patch, I added comments about the rest of the changes
needed to remove variable metadata slot size.

Change-Id: I86c21609d42b902a5b5aac221c1155f1bc293628
Signed-off-by: Nir Soffer <nsoffer@redhat.com>