blockSD: Complete removal of variable slot size
Complete the removal of variable volume metadata slot size. Remove the
slotSize argument to acquireVolumeMetadataSlot() and the single value
VOLUME_MDNUMBLKS, and change the return value of _getOccupiedSlots() to
return only the slots numbers.

This does not change the behavior of the system since we always
allocated one block per slot.

Change-Id: I98754255b9f9ed81250208322567d14cfe4eaee0
Signed-off-by: Nir Soffer <nsoffer@redhat.com>