py3: Fix resourcemanager tests with python 3
The test module was not importable before because we the StringIO module
was removed in python 3. Replacing with six.StringIO allows running the
tests with python 3.

About half of the tests were failing because types.MethodType() has a
different signature in python 3. Using six.create_bound_method() fixes
this issue.

Some tests failed because we tried to call hash() on an object which is
not hashable in python 3. This call was used a hack to silence pyflakes.
Removing the hash call fix the issue and keeps flake8 happy.

Change-Id: I511fbc8a7962bf4e12753be873daaf3cb6d1bf55
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>