py3: tool: Remove 'execCmd' usage in 'passwd' configurator
Substituting usage of deprecated 'commands.execCmd' with 'commands.run'.

'SASL_USERNAME' is defined as a string literal, so it will be bytes in
py2 and unicode in py3. To check its presence in command's output
in 'passwd_isconfigured', we have to explicitly encode it to bytes.
This will give us bytes output for both py2 and py3.

'commands.run's 'input' argument should be bytes, but 'libvirt_password'
returns unicode for both py2 and py3, so again, encoding to bytes
is needed.

Change-Id: I9601b315f04a5bf377bc23463e32614132de4663
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>