py3: Don't use sys.maxint
The HTTP server limits TaskQueue size by sys.maxint.  That constant no
longer exists in Python 3 and doesn't make sense to be used as the
limit anyway.  If we wanted to limit queue size by architecture or so,
six.MAXSIZE would be a better constant and the limit should be used
directly in TaskQueue.

Actually, this server is used only for OVF uploads and downloads,
which are very rare.

The queue is used as a temporary place to hold incoming connections
received by the protocol detector socket.  We create new thread per
connection immediately when new connection arrives, so the size of
this queue is likely to be 0 most of the time, and 1 for few
milliseconds after a new connection is accepted.

If this queue has 10 waiting connections there is something very wrong
with this host, or the server thread taking the task from this queue
crashed, or someone is doing DDOS on Vdsm.  In this case it does not
help anyone to queue the connection.

So let's set the limit to 10.

Change-Id: I946b82fd4a8a99c9e519919840b30927ec33be9d
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>