systemd: Remove systemd_run() supervdsm service
Providing systemd_run() in supervdsm can be used to run any command as
any user. This is insecure and confuses developers. systemd.run() should
be used only by code already running as root, it is not a mechanism to
escalate privileges.

This service was used only in storage.nbd for starting a transient NBD
service. Replace with nbd_start_transient_service() starting qemu-nbd
service.

Change-Id: I4d7de70caad2b48c6edd9bf4566ca0fc3a3418d9
Signed-off-by: Nir Soffer <nsoffer@redhat.com>