tests: Add option to use known temporary directory
When tests depends on running supervdsm we cannot patch supervdsm during
the test. It both code running in the test and in supervdsm need to use
the same path, we need a known temporary directory.

Add new temp_dir() context manager accepting a path argument; if
specified, the temporary directory is created at that path. For safety,
path must not exist. If the temporary directory was created, it will be
deleted at the end of the test.

Use temp_dir in fake_env() instead of testlib.namedTemporaryDir(), and
add data_center argument to control the data center directory when
creating a fake environment. If data_center is not specified we use the
previous behavior and create a new temporary directory for every test.

Change-Id: I27eb71ec681789c07af3b115cbf27a60d62bac1f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>