supervdsm: Override vdsm user and group
Previously we had --sock-user and --sock-group to override the
permission on supervdsm socket. But when running nbd, we need to start
nbd server using vdsm user and group. During the tests, we want to make
it possible to override both the socket permissions and VDSM_USER and
VDSM_GROUP constants.

Change the command line options to --user and --group, and override
the constants.

Change-Id: I649435f2af92865ea861fa65a878551febd3ec28
Signed-off-by: Nir Soffer <nsoffer@redhat.com>