supervdsm: Allow overriding /rhev/data-center
When running nbd server, supervdsm must verify that the served path is
inside the storage repository. To test this, we need to use the same
storage repository in both supervdsm and the test.

Add --data-center argument to control the storage repository location.
Using this option overrides storage.constants REPO_DATA_CENTER and
REPO_MOUNT_DIR.

Change-Id: Ia24a3be03b18cbbbac7b43050d12573d25521294
Signed-off-by: Nir Soffer <nsoffer@redhat.com>