lvm: Support creating LV on specific PV
Add device argument to createLV() ensuring that a new LV is created on
specific PV. This will be used to ensure we create the metadata LV on
the same PV used for LVM metadata.

When removing a device from a storage domain (SDM.move_domain_device,
SDM.reduce_domain), we block removal of the first PV used for LVM
metadata, and the device where the metadata LV is located. Keeping the
metadata LV on the first device ensure that we limit only one device.

Having the metadata LV on the first PV is also required for older
storage domain versions that used PV mapping to set the metadata slot,
instead of the MD_N tags.

Change-Id: I9294cacff5385e489c5d250411c47cb00fc867ea
Signed-off-by: Nir Soffer <nsoffer@redhat.com>