blockSD: Use constant size for the metadata LV
We had complex code calculating the metadata LV size, using minimum size
of 512 MiB. When we extended or resize a PV, we recalculated the
metadata LV size, and extended it if the new size exceed the current
size.

There is no documentation explaining these size calculation. Based on
the code, it looks like we tried to support a use case when we have many
tiny volumes, each using one extent (128 MiB). This use case is not
relevant since we have hard limit of 1947 volumes. Theoretically we can
raise the limit by extending the leases volume, but this was never
implemented.

The actual space needed is less then 1 MiB in V4 (using 512 bytes per
volume), and less than 17 MiB in V5 (using 8 KiB per volume). See the
comment in blockSD for the details.

Simplify the code to use constant size (128 MiB). This is big enough for
16,255 volumes per VG using V5 format. The code calculating the metadata
size and extending the metadata LV was removed.

Change-Id: I2d455cf90ecedb0995e1fdc58e86e0d01372e717
Signed-off-by: Nir Soffer <nsoffer@redhat.com>