vm: Create aligned memory metadata volumes
When creating memory metadata volumes, we create a 10 KiB volume, and
opened the file for update, so writing smaller payload would keep the
file aligned. However it seems that the amount of data we write now is
more than 10 KiB, so we created again unaligned volume.

Here is an example volume:

$ ls -l 7fec9ccd-c649-4aff-9098-5953cb80f284
-rw-r-----. 1 nsoffer nsoffer 11050 Feb 18 14:49 7fec9ccd-c649-4aff-9098-5953cb80f284

When copying this volume, qemu-img segfaults:

$ qemu-img convert -f raw -O raw -p -t none -T none 7fec9ccd-c649-4aff-9098-5953cb80f284 7fec9ccd-c649-4aff-9098-5953cb80f284-copy -o preallocation=falloc
qemu-img: block/io.c:2134: bdrv_co_block_status: Assertion `*pnum && (((*pnum) % (align)) == 0) && align > offset - aligned_offset' failed.
Aborted (core dumped)

This is probably a new qemu bug, but we should not create such volumes.
Change the code to ensure that the volume is aligned.

Change-Id: I2c478b58c58dc3015f134e970de86e6e58e43d4a
Bug-Url: https://bugzilla.redhat.com/1678373
Signed-off-by: Nir Soffer <nsoffer@redhat.com>